+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV57XWZP2WHSFS0XNPEXQ43Q"
+        "01KV5DTNB5WAJ35N965FZ29ET8"
     )
     r"""The unique identifier of the data source"""