+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV57XWZHFGHNJ1Q79JA4R2BM"
+        "01KV5DTNB2QZX1PXBYE0HBTZNC"
     )
     r"""The unique identifier of the data source"""