+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV57XWZM8CEYVX4WW6E1DPQT"
+        "01KV5DTNB4RBGMQ4DCENF8ZJST"
     )
     r"""The unique identifier of the data source"""